Finalize `PermutationCoordinateDescent`, cleanup
