+++ rebasin/__init__.py
@@ -1,6 +1,7 @@
 import rebasin.interpolation
 import rebasin.modules
 import rebasin.paths
+import rebasin.permutation_initializer
 import rebasin.structs
 import rebasin.utils
 from rebasin.permutation_coordinate_descent import PermutationCoordinateDescent
